+++ NEWS.md
@@ -1,6 +1,7 @@
 # HDO.db 1.0.0
 + update data (2024-07-23)
 + delete parse_obo.r and use ontologyIndex::get_ontology to parse obo file
++ add HDOGENE and HDOGENENCG
 
 # HDO.db 0.99.0